The last update unintentionally created a few bugs; bugs now fixed. Added some minor features (see Details).
Details: Last update unintentionally created a few bugs; bugs now fixed. Also added the `verbose` argument to `import_as`, as the function now gives info on conflicts and so on in its attributes. Created some fake packages JUST for the sake of thorough testing. The tests are not part of R CMD CHECK (because it doesn't like packages defined inside package folder), and not uploaded here on GitHub; but I do perform the test with these fake packages everytime I update this R package.